+++ sparrow/src/Classes/SPImage.h
@@ -60,13 +60,13 @@ NS_ASSUME_NONNULL_BEGIN
 /// -------------
 
 /// Sets the texture coordinates of a vertex. Coordinates are in the range [0, 1].
-- (void)setTexCoords:(SPPoint *)coords ofVertex:(int)vertexID;
+- (void)setTexCoords:(SPPoint *)coords ofVertex:(NSInteger)vertexID;
 
 /// Sets the texture coordinates of a vertex. Coordinates are in the range [0, 1].
-- (void)setTexCoordsWithX:(float)x y:(float)y ofVertex:(int)vertexID;
+- (void)setTexCoordsWithX:(float)x y:(float)y ofVertex:(NSInteger)vertexID;
 
 /// Gets the texture coordinates of a vertex.
-- (SPPoint *)texCoordsOfVertex:(int)vertexID;
+- (SPPoint *)texCoordsOfVertex:(NSInteger)vertexID;
 
 /// Readjusts the dimensions of the image according to its current texture. Call this method 
 /// to synchronize image and texture size after assigning a texture with a different size.
