+++ sparrow/src/Classes/SPSprite.m
@@ -134,7 +134,7 @@
     if (_flattenedContents)
     {
         [support finishQuadBatch];
-        [support addDrawCalls:(int)_flattenedContents.count];
+        [support addDrawCalls:_flattenedContents.count];
 
         SPMatrix3D *mvpMatrix = support.mvpMatrix3D;
         float alpha = support.alpha;
